Prepare new FrontDoor CDN infrastructure
Our old CDN, which was originally Verizion and is now called Edgio, is
apparently going to go away in a week, and starting tomorrow Azure is
going to auto-migrate people to new CDN infra. Here we set up a new CDN
based on Azure's FrontDoor, which is now the preferred/only CDN
offering.

These changes set up new endpoints with new custom domains using the new
system. Based on my testing, they seem to work. Next I will turn off the
old custom domains and migrate them over to the new system. This will
necessarily involve a bit of downtime, but it shouldn't be much, and I
don't see a way to avoid that that doesn't involve its own extra layers
of risky complexity. (E.g., the official Azure migration instructions
suggest deploying Azure Traffic Manager temporarily ... that does not
seem like a net win, for us.)

I'll start with the "stacks" domains that are much less
mission-critical. The aim is to make one Git commit for each `terraform
apply` in the migration dance.